warnings
git-svn-id: https://dev.csoft.net/agar/trunk@6576 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ gui/graph.c
@@ -726,7 +726,7 @@ PlaceVertex(AG_Graph *gf, AG_GraphVertex *vtx, AG_GraphVertex **vSorted,
 				PlaceVertex(gf, oVtx, vSorted, ox, oy);
 				break;
 			}
-			theta += (AG_PI*2.0)/6;
+			theta += (float)(AG_PI*2.0)/6;
 			if (theta >= (AG_PI*2.0)) {
 				r += 64.0;
 			}
